+++ tests/documents/test_document.py
@@ -32,19 +32,7 @@ def test_responseWrapper_plate(dummy_file_input):
 def test_responseWrapper_invoice(dummy_file_input):
     response = json.load(open('./tests/data/invoices/v2/invoice.json'))
     parsed_invoice = Response.format_response(response, "invoice", dummy_file_input)
-    assert parsed_invoice.invoice.invoice_date.value == "2020-02-17"
-
-
-from mindee import Invoice
-
-json_repsonse = json.load(open("./tests/data/invoices/v2/invoice.json"))
-invoice_0 = Invoice(json_repsonse["predictions"][0])
-invoice_1 = Invoice(json_repsonse["predictions"][1])
-
-
-def test_merge_pages():
-    merge_document = Document.merge_pages([invoice_0, invoice_1])
-    assert merge_document.invoice_date.value == "2020-02-17"
+    assert parsed_invoice.invoice.invoice_date.value == '2018-09-25'
 
 
 # Receipt tests
@@ -66,5 +54,5 @@ def test_responseWrapper_financial_document_with_receipt(dummy_file_input):
 def test_responseWrapper_financial_document_with_invoice(dummy_file_input):
     response = json.load(open('./tests/data/invoices/v2/invoice.json'))
     parsed_financial_doc = Response.format_response(response, "financial_document", dummy_file_input)
-    assert parsed_financial_doc.financial_document.date.value == "2020-02-17"
+    assert parsed_financial_doc.financial_document.date.value == '2018-09-25'