+++ tests/test_plots.py
@@ -1,0 +1,8 @@
+from mindee.plots import plot_metrics
+
+
+def test_plots():
+    try:
+        plot_metrics(["test", "test2"], [100, 90], [90, 87], './', savefig=False)
+    except:
+        assert False, "Plot raised Exception"